Extrude bot db method to outer file
